+++ cognite/client/data_classes/transformations.py
@@ -2,79 +2,68 @@ from cognite.client.data_classes._base import *
 
 
 class TransformationDestination:
-    """TransformationDestination has static methods to define the target resource type of a transformation
-    """
+    """TransformationDestination has static methods to define the target resource type of a transformation"""
 
     def __init__(self, type: str = None):
         self.type = type
 
     @staticmethod
     def assets():
-        """`To be used when the transformation is meant to produce assets.`_
-        """
+        """To be used when the transformation is meant to produce assets."""
         return TransformationDestination("assets")
 
     @staticmethod
     def timeseries():
-        """`To be used when the transformation is meant to produce time series.`_
-        """
+        """To be used when the transformation is meant to produce time series."""
         return TransformationDestination("timeseries")
 
     @staticmethod
     def assethierarchy():
-        """`To be used when the transformation is meant to produce asset hierarchies.`_
-        """
+        """To be used when the transformation is meant to produce asset hierarchies."""
         return TransformationDestination("assethierarchy")
 
     @staticmethod
     def events():
-        """`To be used when the transformation is meant to produce events.`_
-        """
+        """To be used when the transformation is meant to produce events."""
         return TransformationDestination("events")
 
     @staticmethod
     def datapoints():
-        """`To be used when the transformation is meant to produce numeric data points.`_
-        """
+        """To be used when the transformation is meant to produce numeric data points."""
         return TransformationDestination("datapoints")
 
     @staticmethod
     def stringdatapoints():
-        """`To be used when the transformation is meant to produce string data points.`_
-        """
+        """To be used when the transformation is meant to produce string data points."""
         return TransformationDestination("stringdatapoints")
 
     @staticmethod
     def sequences():
-        """`To be used when the transformation is meant to produce sequences.`_
-        """
+        """To be used when the transformation is meant to produce sequences."""
         return TransformationDestination("sequences")
 
     @staticmethod
     def files():
-        """`To be used when the transformation is meant to produce files.`_
-        """
+        """To be used when the transformation is meant to produce files."""
         return TransformationDestination("files")
 
     @staticmethod
     def labels():
-        """`To be used when the transformation is meant to produce labels.`_
-        """
+        """To be used when the transformation is meant to produce labels."""
         return TransformationDestination("labels")
 
     @staticmethod
     def relationships():
-        """`To be used when the transformation is meant to produce relationships.`_
-        """
+        """To be used when the transformation is meant to produce relationships."""
         return TransformationDestination("relationships")
 
     @staticmethod
     def raw(database: str = "", table: str = ""):
-        """`To be used when the transformation is meant to produce raw table rows.`_
+        """To be used when the transformation is meant to produce raw table rows.
 
         Args:
-            database (str) – database name of the target raw table.
-            table (str) – name of the target raw table
+            database (str): database name of the target raw table.
+            table (str): name of the target raw table
 
         Returns:
             TransformationDestination pointing to the target table
@@ -119,23 +108,24 @@ class Transformation(CogniteResource):
     Args:
         id (int): A server-generated ID for the object.
         external_id (str): The external ID provided by the client. Must be unique for the resource type.
-        name (str): The name of the ExtractionPipeline.
-        description (str): The description of the ExtractionPipeline.
-        data_set_id (int): The id of the dataset this ExtractionPipeline related with.
-        raw_tables (List[Dict[str, str]): list of raw tables in list format: [{"dbName": "value", "tableName" : "value"}].
-        last_success (int): Milliseconds value of last success status.
-        last_failure (int): Milliseconds value of last failure status.
-        last_message (str): Message of last failure.
-        last_seen (int): Milliseconds value of last seen status.
-        schedule (str): undefined/triggered/streamed/cron regex.
-        contacts (List[Dict[str, Any]]): list of contacts [{"name": "value", "email": "value", "role": "value", "sendNotification": boolean},...]
-        metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value. Limits: Maximum length of key is 128 bytes, value 10240 bytes, up to 256 key-value pairs, of total size at most 10240.
-        source (str): Source text value for ExtractionPipeline.
-        documentation (str): Documentation text value for ExtractionPipeline.
+        name (str): The name of the Transformation.
+        query (str): SQL query of the transformation.
+        destination (TransformationDestination): see TransformationDestination for options.
+        conflict_mode (str): What to do in case of id collisions: either "abort", "upsert", "update" or "delete"
+        is_public (bool): Indicates if the transformation is visible to all in project or only to the owner.
+        ignore_null_fields (bool): Indicates how null values are handled on updates: ignore or set null.
+        source_api_key (str): Configures the transformation to authenticate with the given api key on the source.
+        destination_api_key (str): Configures the transformation to authenticate with the given api key on the destination.
+        source_oidc_credentials (Optional[OidcCredentials]): Configures the transformation to authenticate with the given oidc credentials key on the destination.
+        destination_oidc_credentials (Optional[OidcCredentials]): Configures the transformation to authenticate with the given oidc credentials on the destination.
         created_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
         last_updated_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
-        created_by (str): ExtractionPipeline creator, usually email.
-        skip_notifications_in_minutes (int): Number value for system to skip sending email notification in minutes after last sending.
+        owner (str): Owner of the transformation: requester's identity.
+        owner_is_current_user (bool): Indicates if the transformation belongs to the current user.
+        has_source_api_key (bool): Indicates if the transformation is configured with a source api key.
+        has_destination_api_key (bool): Indicates if the transformation is configured with a destination api key.
+        has_source_oidc_credentials (bool): Indicates if the transformation is configured with a source oidc credentials set.
+        has_destination_oidc_credentials (bool): Indicates if the transformation is configured with a destination oidc credentials set.
         cognite_client (CogniteClient): The client to associate with this object.
     """