+++ CHANGELOG.md
@@ -14,6 +14,12 @@ Changes are grouped as follows
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+
+## [2.35.0] - 2021-11-22
+### Added
+- `transformations` api client, which allows the creation, deletion, update and retrieval of transformations.
+- `transformations.schedules` api client, which allows the schedule, unschedule and retrieval of recurring runs of a transformation.
+
 ## [2.34.0] - 2021-11-5
 ### Added
 - Added support for `data_set_id` on template views